fix enduser run

+++ src/agilab/core/agi-env/src/agi_env/agi_env.py
@@ -845,15 +845,14 @@ class AgiEnv:
         self.gitignore_file = self.app_abs / ".gitignore"
         dest = AgiEnv.resources_path
         src = self.agilab_src / "resources"
-        for file in src.iterdir():
-            if not file.is_file():
-                continue
-
-            dest_file = dest / file.name
-            if dest_file.exists():
-                continue
-
-            shutil.copy2(file, dest_file)
+        if src.exists():
+            for file in src.iterdir():
+                if not file.is_file():
+                    continue
+                dest_file = dest / file.name
+                if dest_file.exists():
+                    continue
+                shutil.copy2(file, dest_file)
         # shutil.copytree(self.agilab_src / "resources", dest, dirs_exist_ok=True)
